+++ src/components/client/about/helpers/managementBoardData.tsx
