+++ src/components/client/about/helpers/principlesData.tsx
