+++ src/components/client/about/icons/AwarenessIcon.tsx
