+++ src/components/client/about/icons/ExpertiseIcon.tsx
