+++ src/components/client/about/icons/PrivacyIcon.tsx
