+++ src/components/client/about/icons/TransparencyIcon.tsx
