+++ src/components/client/about/icons/VoluntaryIcon.tsx
