+++ src/components/client/about/sections/AboutTheTeamSection/AboutTheTeamSection.styled.tsx
