+++ src/components/client/about/sections/ManagementBoardSection/ManagementBoardSection.tsx
@@ -2,37 +2,45 @@ import { useTranslation } from 'next-i18next'
 
 import { Grid } from '@mui/material'
 
-import { data } from '../../helpers/associationMembersData'
+import { data } from '../../helpers/managementBoardData'
 
-import { TeamMemberWrapper } from './AssociationMembersSection.styled'
 import {
-  AboutHeading,
+  ManagementBoardHeading,
+  ТeamMemberWrapper,
+  LinkedInButton,
+} from './ManagementBoardsection.styled'
+import {
   AboutWrapper,
   Avatar,
-  LinkedInButton,
+  TeamMemberName,
   LinkedInIcon,
   LinkedInText,
-  TeamMemberName,
-} from 'components/about/AboutPage.styled'
+} from 'components/client/about/AboutPage.styled'
+import { TeamMemberDescription } from './TeamMemberDescription'
 
-export default function AssociationMembersSection() {
+export default function ManagementBoardSection() {
   const { t } = useTranslation('about')
 
   return (
     <Grid component="section">
-      <AboutHeading variant="h4">{t('about.association-members')}</AboutHeading>
+      <ManagementBoardHeading variant="h4">
+        {t('about.management-board-members')}
+      </ManagementBoardHeading>
       <AboutWrapper>
         {data.map((teamMember) => (
-          <TeamMemberWrapper key={teamMember.name}>
+          <ТeamMemberWrapper key={teamMember.name}>
             <Avatar alt="Team member avatar" src={teamMember.img} width={200} height={200} />
-            <TeamMemberName variant="subtitle1">{teamMember.name}</TeamMemberName>
+            <TeamMemberName variant="subtitle1" height={50}>
+              {teamMember.name}
+            </TeamMemberName>
+            <TeamMemberDescription description={teamMember.description} />
             {teamMember.linkedInProfile ? (
               <LinkedInButton href={teamMember.linkedInProfile} target="_blank">
                 <LinkedInIcon color="action" fontSize="large" />
                 <LinkedInText variant="subtitle1">{t('about.linkedIn')}</LinkedInText>
               </LinkedInButton>
             ) : null}
-          </TeamMemberWrapper>
+          </ТeamMemberWrapper>
         ))}
       </AboutWrapper>
     </Grid>
