+++ src/components/client/about/sections/ManagementBoardSection/ManagementBoardsection.styled.tsx
