+++ src/components/client/about/sections/ManagementBoardSection/TeamMemberDescription.tsx
