+++ src/components/client/about/sections/PrinciplesThatUniteUsSection/PrincipleCard.styled.tsx
