+++ src/components/client/about/sections/SupervisoryBoardSection/SupervisoryBoardSection.styled.tsx
